+++ nueve4/template-parts/content-download-single.php
@@ -6,7 +6,7 @@
  * @package Nueve4
  */
 
-if ( defined( 'NEVE_PRO_PATH' ) && get_option( 'nv_pro_easy_digital_downloads_status' ) ) {
+if ( defined( 'NUEVE4_PRO_PATH' ) && get_option( 'nv_pro_easy_digital_downloads_status' ) ) {
 	do_action( 'nueve4_do_single_download', 'single-download' );
 } else {
 	do_action( 'nueve4_do_single_post', 'single-post' );